Start the development of the search module in the server
This change introduces two new created classes for the search mechanism.
One of those classes is the controller that gives the route to get the
information needed. This class uses the second class which is the
handler responsible to make the search action itself.

For now, the search is hardcoded to certain values. This helps to test
the mechanism. Next development necessary for this module is the
creation of the model which receives the information related to the
search from the user in order to customize the search.